# server/openapi.yaml
openapi: 3.1.0
info:
  title: evcc
  description: Solar charging. Super simple.
  version: "0.2.0"
  contact:
    url: https://github.com/evcc-io/evcc
servers:
  - url: https://demo.evcc.io/api
...
paths:
...
  /batterydischargemode/{batteryDischargeMode}:
    post:
      operationId: setBatteryDischargeMode
      summary: Set battery support during fast and planned charging
      description: "Allow unrestricted home battery support, use it down to the configured reserve SoC, or prevent it."
      externalDocs:
        url: https://docs.evcc.io/en/features/battery
      tags:
        - battery
      parameters:
        - $ref: "#/components/parameters/batteryDischargeMode"
      responses:
        "200":
          $ref: "#/components/responses/BatteryDischargeModeResult"
...
  /db/reset:
    post:
      operationId: resetDatabase
      summary: Reset database
      description: "Selectively deletes sessions and/or settings from the database. Session users must supply the admin password in the X-Admin-Password header. API key holders via Bearer token are exempt. The instance restarts after a successful reset."
      tags:
        - db
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/adminPassword"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessions:
                  type: boolean
                  description: Delete all charging sessions
                settings:
                  type: boolean
                  description: Delete all settings, configs, and meters
      responses:
        "204":
          description: Reset successful, instance is restarting
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  schemas:
    BatteryDischargeMode:
      description: Home battery support during fast and planned charging
      type: string
      example: reserve
      enum:
        - allow
        - reserve
        - prevent
...
    Password:
      description: Admin password
      type: string
...
  parameters:
    adminPassword:
      name: X-Admin-Password
      in: header
      description: Admin password (required for session auth, not needed for API key)
      schema:
        $ref: "#/components/schemas/Password"
...
    batteryDischargeMode:
      name: batteryDischargeMode
      description: Home battery support during fast and planned charging
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/BatteryDischargeMode"
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Login and try again
      content:
        text/plain:
          schema:
            type: string
            enum:
              - Unauthorized
...
    BatteryDischargeModeResult:
      description: Home battery support mode
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BatteryDischargeMode"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Long-lived API key (prefixed `evcc_`, managed via /auth/apikey).
        Grants access to /api/db/* without requiring the X-Admin-Password header.